requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv
